requirements.txt
pycryptodome==3.19.0
websockets==12.0